+++ src/processTargets/marks/CursorStage.ts
@@ -3,7 +3,7 @@ import type { CursorMark } from "../../typings/targetDescriptor.types";
 import { ProcessedTargetsContext } from "../../typings/Types";
 import { isReversed } from "../../util/selectionUtils";
 import { MarkStage } from "../PipelineStages.types";
-import UntypedRangeTarget from "../targets/UntypedRangeTarget";
+import UntypedTarget from "../targets/UntypedTarget";
 
 export default class CursorStage implements MarkStage {
   constructor(private modifier: CursorMark) {}
@@ -11,7 +11,7 @@ export default class CursorStage implements MarkStage {
   run(context: ProcessedTargetsContext): Target[] {
     return context.currentSelections.map(
       (selection) =>
-        new UntypedRangeTarget({
+        new UntypedTarget({
           editor: selection.editor,
           isReversed: isReversed(selection.selection),
           contentRange: selection.selection,
