+++ src/processTargets/targets/index.ts
@@ -22,5 +22,5 @@ export * from "./SurroundingPairTarget";
 export { default as SurroundingPairTarget } from "./SurroundingPairTarget";
 export * from "./TokenTarget";
 export { default as TokenTarget } from "./TokenTarget";
-export * from "./UntypedRangeTarget";
-export { default as UntypedRangeTarget } from "./UntypedRangeTarget";
+export * from "./UntypedTarget";
+export { default as UntypedTarget } from "./UntypedTarget";
